requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.1.1
pymongo==4.3.3
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0.post1
gunicorn==21.2.0 